Done: Create new service
Note the in preview  section we have same component which is used to preview service on client side. Later on this will be loaded using may be some generated id of service. and in the end there is button of publish which means after publish this service will be visible to public